ovl: create directories inside merged parent opaque
The benefit of making directories opaque on creation
is that lookups can stop short when they reach the
original created directory, instead of continue lookup
the entire depth of parent directory stack.

The best case is overlay with N layers, performing
lookup for first level directory, which exists only in upper.
In that case, there will be only one lookup instead of N.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>