fanotify: report filename to sb/inode watch with FS_EVENT_ON_CHILD
With fanotify_init() flag FAN_REPORT_FILENAME, when adding an inode watch
with FS_EVENT_ON_CHILD, deliver events to listener with fid of the parent
and the filename of the child entry, like inotify. In that case, the
FS_EVENT_ON_CHILD flag is included in the event's mask.

When adding a sb watch with FS_EVENT_ON_CHILD, deliver some events to
listener in two flavors. Once with the child fid and without filename and
once with the parent fid and with the child entry filename.

This is only relevant to events open/modify/attrib/close, which can be
reported to both parent and child.
Directory entry modification events (create/detete/move) always include
filename and self events (move_self/delete_self) never include filename.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>