ovl: copy file to snapshot on open for write
On open for write of file in a snapshot fs mount, call ovl_copy_up() on
the snapshot overlay dentry, which will copy the file to the snapshot
overlay's upper dir if needed.

Then, ovl_open() will open a file with its own upper dentry (not the
snapshot overlay's upper dentry), so the stable copy of the file will
remain in the snapshot overlay's upper dir.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>