ovl: remount snapshot mount with new snapshot overlay
Allow changing the snapshot overlay of snapshot mount by
provding a new snapshot=<snappath> argument to remount.

ovl_snapshot_remount() changes the value of __snapmnt
to the new snapshot overlay and the value of __snapdentry
of the snapshot mount root entry to the snapshot overlay root.

The rest of the snapshot mount dentries that reference old
snapshot overlay dentries, will be invalidated lazily on
the next lookup by ovl_snapshot_revalidate().

Signed-off-by: Amir Goldstein <amir73il@gmail.com>