ovl: implement nfs_export=nested
For nested overlayfs with single lower layer, nfs_export=nested encodes
lower file handles that are the same as the file handles encoded directly
from the real nested lower fs for the same objects.

For an upper file handle from the lower nested overlayfs, we set a
'nested' flag in ovl_fh header, so we know to distinguish between real
upper file handle that should be decoded from real upper layer and
nested upper file handle that should be decoded from nested lower layer.
For the maximum allowed overlay nesting depth of 1, one bit is enough.

This is going to be used by overlayfs snapshots when rotating an upper
layer to be a nested upper layer (a.k.a previous snapshot).

Signed-off-by: Amir Goldstein <amir73il@gmail.com>