ovl: copy open-for-write file to new snapshot
When a file is open for write and a new snapshot was taken, the copy
to old snapshot is now not uptodate and we need to copy file to new
snapshot on next change.

The snapflags and snapid in dentry minimize the overhead of check
whether file needs to be copied to snapshot on every change.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>