ovl: store snapshot id along side snapshot dentry flags
Snapshot fs dentry flags are use test if dentry needs cow to snapshot.
On snapshot take, those flags need to be invalidated across all dentries.
To achieve that, we store the snapshot id along side the flags and along
side the snapshot and we test that flags refer to the current snapshot id.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>