fsnotify: add support for const marks
A constant mark mask cannot be modified - bits cannot be added nor
removed from mask or ignored mask.

The only way to remove a const mark is to destroy all marks of its type
with FAN_MARK_FLUSH or close the group fd.

There is currently no support for adding a const mark from user API.
It is going to be set implicitly on an fanotify subtree mark.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>