fanotify: report operation id instead of pid
With fanotify_init() flag FAN_REPORT_OPID, report an operation id
instead of process pid in event.pid.

opid is a rebranding of the inotify rename cookie, which is going to be
extended to couple more fanotify event types that result from the same
fs "operation" - that is, from the same fsnotify hook.

Since reporting pid/tid is mostly required for permission events and
reporting opid is required for dirent events, they are never needed for
the same use case, so it is convetient to repurpose event.pid.

Another way to look at this is as async operations replacement for
pid. If a process issues async fs operations, the original meaning of
event.pid is lost, but we can use opid to replace it with async job id.

FAN_REPORT_OPID requires FAN_REPORT_FID and it is mutualy exclusive
with FAN_REPORT_TID.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>