fanotify: report opid with FAN_MOVE_SELF event
opid, a.k.a cookie, is used in inotify to couple IN_MOVED_FROM and
IN_MOVED_TO events. Those two inotify events also carry the old and new
names, which are not available in respective fanotify events.

We therefore pass the same opid also on FAN_MOVE_SELF event.
The fid reported on this event can be used to get the new file path.
The fids reported on FAN_MOVED_FROM and FAN_MOVED_TO events can be
used to get the old and new parent paths.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>