fanotify: introduce events FAN_LINK_SELF/FAN_UNLINK_SELF
Akin to FAN_MOVE_SELF, those events are generated with same opid
as FAN_CREATE/FAN_DELETE, but with fid of the victim child.

These events are not the same as FAN_DELETE_SELF, which is generated
when last reference on inode drops.

The aliases FAN_LINK/FAN_UNLINK were created to FAN_CREATE/FAN_DELETE.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>