ovl: prepare internal helper for snapshot lookup
ovl_lookup_real() is going to be use to find out if lower object needs
to be copied to snapshot, so make it non-static.

A directory whited out in snapshot without being indexed is a common case
with overlay snapshot, for example on mkdir in lower dir.

Do not warn about this case when looking up overlay snapshot dir by lower
dir inode.  Convert EIO error to ESTALE in this case, to indicate that
the upper object is not a copy up of the lower object.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>