ovl: suppress lookup by real warnings from snapshot lookup
ovl_lookup_real() is going to be called for snapshot lookup, so make it
non-static.

A directory whited out in snapshot without being indexed is a common
case, for example on mkdir in snapshot fs.

Do not warn about it when looking up snapshot dir by upper dir.
Convert EIO error to ESTALE, because snapshot mount needs to treat this
condition that same as a whited out directory.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>