ovl: prevent taking a snapshot if writable maps to files exist
Overlayfs and snapshot fs do not implement address space ops so we cannot
copy file to snapshot on first page fault.

To prevent writing into writable memory maps without copy to snapshot,
we need to prevent taking a snapshot if writable shared maps exist on
upper fs real files.

Snapshot is taken either by freeze fs or by remount read-only.
Remount read-only is prevented by elevated mnt_writers from the reference
of "realfile" that is used to create the mapping. "realfile" has the
underlying inode, but the overlay path/mnt.

Freeze fs is prevented by a new accounting of writable shared maps in the
filesystem. We also deny creating new shared writable maps while
filesystem is frozen.

Because we do not have an unmap() hook, we use the release() hook to
detect that a file that used to have writable maps does not have
writable maps at the time it is being released - to drop the shared
writable maps count.

With this mechanism, a process may create a writable map to a file,
close the file descriptor and exit, leaving the writable maps count
elevated and preventing to take snapshots. That is the necessary evil
we chose as an interim solution until overlayfs/snapshot fs implement
address space operations and support proper freeze for page faults.

On the other side, we deny creating new writable maps if filesystem is
frozen or if an open for write file needs to be copied to snapshot.
In that case, user can re-open the file for write and then create the map.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>