fsnotify: pass dentry instead of inode for events where possible
Define a new data type to pass for event - FSNOTIFY_EVENT_DENTRY
and use it to pass the dentry instead of it's ->d_inode for events where
the dentry is available.

The special case of fsnotify_link() hook has both dentry and inode
which usually obay to d_inode(dentry) == inode, so only dentry can be
passed on, but the comment says that "some" filesytems leave d_inode
NULL. Check that unlikely() case and fall back to passing the inode.

Soon, we are going to use the dentry data type to check ancestors for
subtree marks.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>