vfs: optimize some ancestry walks in lock_rename()
Containers tend to increase the depth of filesystem directory trees and
the nested container case, increases them even more.

lock_rename() walks source and target dir ancestries up to filesystem
root. We can optimize a common case, where source and target dirs have a
common ancestor mount point (e.g. bind mount).

On first iteration, we stop at the closet mount point ancestors of p1
and of p2 if they exist.  On second iteration, if stopped at mount point
ancestor, continue walk all the way to filesystem root or to the other
directory.  Both first and second iterations will stop if they find that
the closest mount point ancestor of the other directory is a common mount
point ancestor.

The motivation for this change was not an optimization, but to detect
renames that cross (private) mount points (e.g. overlayfs layer root).
This is going to be implemented by following patches. In the mean while
the optimization doesn't hurt.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>