vfs: add a security hook for renaming across mount points
Containers runtimes often use bind mounts, overlayfs, shiftfs to provide
access to a fraction of a filesystem to containers.

A security policy may wish to prevent moving files and directories in
and out of the contained section of the filesystem.

To that end, LSMs can use the security_inode_rename() and
security_path_rename() hooks, but it would be inefficient to check
the ancestors of the source and target for every rename in the
filesystem.

Instead, we add a new hook called security_inode_rename_xmnt(),
which is only called if the rename is about to cross any mount point
(in any mount namespace).  One or two crossed mount point are provided in
the hook arguments, one is ancestor of the source and the other is an
ancestor of the target, but there could be more crossed mount points.
LSM needs to walk the ancestors to find all the crossed mount points.

lock_rename() is extended to lock_rename_xmnt() in order to find the
crossed mount points while doing the same ancestry walks it does already
for figuring out the correct locking order.

The new hook is initially intended to prevent moving objects between
underlying overlayfs layers.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>