ovl: strict upper fs requirements for remote upper fs
Overlayfs works sub-optimally with upper fs that has no
xattr/d_type/O_TMPFILE/RENAME_WHITEOUT support. We should basically
deprecate support for those filesystems, but so far, we only issue a
warning and don't fail the mount for the sake of backward compat.
Some features are already being disabled with no xattr support.

For newly supported remote upper fs, we do not need to worry about
backward compatibility, so we can fail the mount if upper fs is a
sub-optimal filesystem.

This reduces the in-tree remote filesystems supported as upper to
just FUSE, for which the remote upper fs support was added.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>