ovl: use a private non-persistent ino pool
There is no reason to deplete the system's global get_next_ino() pool
for overlay non-persistent inode numbers and there is no reason at all
to allocate non-persistent inode numbers for non-directories.

For non-directories, it is much better to leave i_ino the same as
real i_ino, to be consistent with st_ino/d_ino.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>