ovl: index dir act as work dir
With index=on, let index dir act as the work dir for copy up and
cleanups.  This will help implementing whiteout inode sharing.

We still create the "work" dir on mount regardless of index=on
and it is used to test the features supported by upper fs.
One reason is that before the feature tests, we do not know if
index could be enabled or not.

The reason we do not use "index" directory also as workdir with
index=off is because the existence of the "index" directory acts
as a simple persistent signal that index was enabled on this
filesystem and tools may want to use that signal.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>