ovl: optional sync of upper directory after copy up
With regular overlay copy up, fsync is done in ovl_copy_up_data().
With overlay snapshot copy up, only an empty file or its parent dir are
created as change markers.

We need to make sure that the change markers are persistent on-disk
before making modifications to make sure that after a crash the change
markers will be found.

Add the option to request copy up to fsync the upper dir before marking
the overlay dentry copied up.  fsync of upper dir is requested by passing
in the flag combination O_SYNC|O_DIRECTORY to ovl_copy_up_flags().

This flag combination is not possible from userspace when opening a file
for write, so this functionality is only internal for now.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>