ovl: sync directory after copy to metacopy snapshot
With regular snapshot, fsync is done in ovl_copy_up_data() on copy of file
to snapshot before writing new file data to disk. The file fsync implies
making the parent dir persistent, so there is no risk of not recording
the modification after a crash.

With metacopy snapshot, file is not copied up, only its parent dir, so we
need to make sure that copied up parent is persistent on-disk before writing
new file data to disk to make sure that after a crash the modification is
recorded.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>