ovl: implement lookup in change tracking snapshot
Snapshot lookup is about finding out if a lower path has already been
copied up or whited out in the change tracking snapshot.  This is done
in a similar way of decoding a lower nfs file handle of an overlay
mount without directory inode index entires and redirects.

Before modification of lower layer path, we look for a dentry in the
overlay snapshot whose lower is the underlying lower path.

When looking up in overlay snapshot, what we need to know about the
overlay snapshot dentry is if it is positive/upper/opaque so we know if
it the lower object was already copied up or whited out in snapshot.

We rely on snapshot lookup to populate the dentry and inode cache of the
overlay filesystem, so we can quickly find the overlay snapshot dentry
by the underlying dentry on subsequent calls.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>