ovl: create markers in change tracking snapshot before modify
On FS_PRE_MODIFY_NAME event, if lower path is positive and not already
copied to snapshot, call ovl_copy_up_flags(O_TRUNC) on the overlay
snapshot dentry to create the directory ancestry and maybe create an
empty file as a "change marker" in overlay snapshot upper dir if needed.

On FS_PRE_MODIFY event, only the directory ancestry is created in the
overlay snapshot upper dir if needed.

A directory entry in the overlay snapshot upper tree indicates that there
may be a modification somewhere in the lower tree under that entry.

A non-directory entry in the overlay snapshot upper tree indicates that
there may be modified or new objects (possibly a new directory tree) in
the lower path under that entry.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>